Extract MovableControl
